Remove standalone mode
