+++ cri.go
@@ -17,6 +17,7 @@ limitations under the License.
 package cri
 
 import (
+	"flag"
 	"path/filepath"
 
 	"github.com/containerd/containerd/log"
@@ -24,8 +25,9 @@ import (
 	"github.com/containerd/containerd/plugin"
 	imagespec "github.com/opencontainers/image-spec/specs-go/v1"
 	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
 
-	"github.com/containerd/cri-containerd/cmd/cri-containerd/options"
+	criconfig "github.com/containerd/cri-containerd/pkg/config"
 	"github.com/containerd/cri-containerd/pkg/server"
 )
 
@@ -35,7 +37,7 @@ const criVersion = "v1alpha2"
 // TODO(random-liu): Use github.com/pkg/errors for our errors.
 // Register CRI service plugin
 func init() {
-	config := options.DefaultConfig().PluginConfig
+	config := criconfig.DefaultConfig()
 	plugin.Register(&plugin.Registration{
 		Type:   plugin.GRPCPlugin,
 		ID:     "cri",
@@ -57,8 +59,8 @@ func initCRIService(ic *plugin.InitContext) (interface{}, error) {
 	ic.Meta.Platforms = []imagespec.Platform{platforms.DefaultSpec()}
 	ic.Meta.Exports = map[string]string{"CRIVersion": criVersion}
 	ctx := ic.Context
-	pluginConfig := ic.Config.(*options.PluginConfig)
-	c := options.Config{
+	pluginConfig := ic.Config.(*criconfig.PluginConfig)
+	c := criconfig.Config{
 		PluginConfig: *pluginConfig,
 		// This is a hack. We assume that containerd root directory
 		// is one level above plugin directory.
@@ -69,6 +71,10 @@ func initCRIService(ic *plugin.InitContext) (interface{}, error) {
 	}
 	log.G(ctx).Infof("Start cri plugin with config %+v", c)
 
+	if err := setGLogLevel(); err != nil {
+		return nil, errors.Wrap(err, "failed to set glog level")
+	}
+
 	s, err := server.NewCRIContainerdService(c)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to create CRI service")
@@ -77,10 +83,32 @@ func initCRIService(ic *plugin.InitContext) (interface{}, error) {
 	// Use a goroutine to initialize cri service. The reason is that currently
 	// cri service requires containerd to be initialize.
 	go func() {
-		if err := s.Run(false); err != nil {
+		if err := s.Run(); err != nil {
 			log.G(ctx).WithError(err).Fatal("Failed to run CRI service")
 		}
 		// TODO(random-liu): Whether and how we can stop containerd.
 	}()
 	return s, nil
+}
+
+// Set glog level.
+func setGLogLevel() error {
+	l := logrus.GetLevel()
+	if err := flag.Set("logtostderr", "true"); err != nil {
+		return err
+	}
+	switch l {
+	case log.TraceLevel:
+		return flag.Set("v", "5")
+	case logrus.DebugLevel:
+		return flag.Set("v", "4")
+	case logrus.InfoLevel:
+		return flag.Set("v", "2")
+	// glog doesn't support following filters. Defaults to v=0.
+	case logrus.WarnLevel:
+	case logrus.ErrorLevel:
+	case logrus.FatalLevel:
+	case logrus.PanicLevel:
+	}
+	return nil
 }
